- Corrigé une plantée liée au nouveau code de gestion des images dans le clipboard; le dossier qui contient les images temporaires peut ne pas exister. - Profité pour utiliser System.IO.Path.Combine à la place de string.Concat. - Profité pour modifier le nom du dossier des temporaires pour qu'il dépende du soft utilisé (Creative Docs .NET et Crésus Documents peuvent s'exécuter en parallèle) et du numéro de processus (s'il y a deux instances de CrDoc).
Note: marrant, j'ai fait presque les mêmes corrections que Daniel dans son dernier commit touchant à Modifier/DocumentEditor !



git-svn-id: https://svn.opac.ch/svn/cr/branches/vs2005/Epsitec.Cresus@11874 787ff09c-d2b5-7842-b444-8dd0a418631d

+++ Epsitec.Cresus/Common.DocumentEditor/DocumentEditor.cs
@@ -5537,11 +5537,27 @@ namespace Epsitec.Common.DocumentEditor
 		protected void DeleteTemporaryFiles()
 		{
 			//	Supprime toutes les images dans le dossier temporaire, issue des opérations "coller".
-			string search = string.Concat(Modifier.ClipboardFilenamePrefix, "*", Modifier.ClipboardFilenamePostfix);
-			string[] list = System.IO.Directory.GetFiles(Modifier.TemporaryDirectory, search);
-			foreach (string filename in list)
+			try
 			{
-				System.IO.File.Delete(filename);
+				if (System.IO.Directory.Exists(Modifier.TemporaryDirectory))
+				{
+					string filter = string.Concat(Modifier.ClipboardFilenamePrefix, "*", Modifier.ClipboardFilenamePostfix);
+					string[] list = System.IO.Directory.GetFiles(Modifier.TemporaryDirectory, filter);
+					foreach (string filename in list)
+					{
+						try
+						{
+							System.IO.File.Delete(filename);
+						}
+						catch
+						{
+						}
+					}
+					System.IO.Directory.Delete(Modifier.TemporaryDirectory);
+				}
+			}
+			catch
+			{
 			}
 		}
